prep work for new requester

+++ provider.go
@@ -45,9 +45,9 @@ func (r DNSRR) RR() dns.RR {
 	// Build an RR header
 	rrhdr := dns.RR_Header{
 		Name:     r.Name,
-		Rrtype:   uint16(r.Type),
+		Rrtype:   r.Type,
 		Class:    dns.ClassINET,
-		Ttl:      uint32(r.TTL),
+		Ttl:      r.TTL,
 		Rdlength: uint16(len(r.Data)),
 	}
